requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
astroid==3.0.1
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
autopep8==2.0.4
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.10.0
blacken-docs==1.16.0
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.4.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
comm==0.1.3
coverage==7.3.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.7
Django==4.2.4
executing==1.2.0
fastapi==0.104.0
fastapi-responses==0.2.1
fastjsonschema==2.18.0
filelock==3.12.4
flake8==6.1.0
fqdn==1.5.1
h11==0.14.0
identify==2.5.30
idna==3.4
iniconfig==2.0.0
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
isort==5.12.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
jupytext==1.15.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.1
mypy==1.6.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.1
nbqa==1.7.0
nbval==0.10.0
nest-asyncio==1.5.7
nodeenv==1.8.0
norminette==3.3.51
notebook==7.0.0
notebook_shim==0.2.3
overrides==7.3.1
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.3.0
pre-commit==3.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.15.1
pylint==3.0.2
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pyupgrade==3.15.0
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
ruff==0.1.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.27.0
terminado==0.17.1
tinycss2==1.2.1
tokenize-rt==5.2.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==2.0.7
uvicorn==0.23.2
virtualenv==20.24.5
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
